Move schema to parser
