+++ src/features/ingestion/domain/__init__.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass
 from functools import cached_property
 
 from src.core import Entity
-from src.features.ingestion.schema import (
+from src.features.ingestion.parser.schema import (
     AcademicBackground,
     GeneralData,
     ProfessionalExperience,
